requirements.txt
streamlit>=1.40
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-community>=0.3.7
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.2
langgraph>=0.2.45
chromadb>=0.5.20
tavily-python>=0.5.0
python-dotenv>=1.0.1
tiktoken>=0.8.0
